Show a DB-unreachable banner on the sign-in page.
Splits the sign-in route into a server component that probes
getDbStatus() + a client form. When Postgres is unreachable (ECONNREFUSED,
missing tables, DATABASE_URL unset), the form renders a destructive
banner explaining the situation, surfaces the raw error for operators,
and disables the inputs + submit button so users don't get an opaque
500 from /api/auth/sign-in/email.
