+++ app/(auth)/sign-in/page.tsx
@@ -1,117 +1,7 @@
-"use client";
+import { getDbStatus } from "@/lib/database/utils";
+import SignInForm from "./sign-in-form";
 
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import {
-	Card,
-	CardContent,
-	CardDescription,
-	CardHeader,
-	CardTitle,
-} from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { authClient } from "@/lib/auth/auth-client";
-
-export default function SignInPage() {
-	const router = useRouter();
-	const [email, setEmail] = useState("");
-	const [password, setPassword] = useState("");
-	const [error, setError] = useState("");
-	const [isLoading, setIsLoading] = useState(false);
-
-	const handleSubmit = async (e: React.FormEvent) => {
-		e.preventDefault();
-		setError("");
-		setIsLoading(true);
-
-		try {
-			const { data, error: authError } = await authClient.signIn.email({
-				email,
-				password,
-			});
-
-			if (authError) {
-				setError(authError.message || "Failed to sign in. Please try again.");
-				setIsLoading(false);
-				return;
-			}
-
-			if (data) {
-				// Redirect to home page on successful sign in
-				router.push("/");
-				router.refresh();
-			}
-		} catch (_err) {
-			setError("An unexpected error occurred. Please try again.");
-			setIsLoading(false);
-		}
-	};
-
-	return (
-		<div className="flex min-h-screen items-center justify-center p-4">
-			<Card className="w-full max-w-md">
-				<CardHeader>
-					<CardTitle className="text-2xl">Sign In</CardTitle>
-					<CardDescription>
-						Enter your email and password to access your account
-					</CardDescription>
-				</CardHeader>
-				<CardContent>
-					<form onSubmit={handleSubmit} className="space-y-4">
-						{error && (
-							<div className="rounded-md bg-destructive/10 p-3 text-sm text-destructive">
-								{error}
-							</div>
-						)}
-						<div className="space-y-2">
-							<Label htmlFor="email">Email</Label>
-							<Input
-								id="email"
-								type="email"
-								placeholder="you@example.com"
-								value={email}
-								onChange={(e) => setEmail(e.target.value)}
-								required
-								autoComplete="email"
-								disabled={isLoading}
-							/>
-						</div>
-						<div className="space-y-2">
-							<div className="flex items-center justify-between">
-								<Label htmlFor="password">Password</Label>
-								<Link
-									href="/recover"
-									className="text-sm text-muted-foreground hover:text-primary"
-								>
-									Forgot password?
-								</Link>
-							</div>
-							<Input
-								id="password"
-								type="password"
-								placeholder="••••••••"
-								value={password}
-								onChange={(e) => setPassword(e.target.value)}
-								required
-								autoComplete="current-password"
-								disabled={isLoading}
-							/>
-						</div>
-						<Button type="submit" className="w-full" disabled={isLoading}>
-							{isLoading ? "Signing in..." : "Sign In"}
-						</Button>
-						<div className="text-center text-sm text-muted-foreground">
-							Don't have an account?{" "}
-							<Link href="/sign-up" className="text-primary hover:underline">
-								Sign up
-							</Link>
-						</div>
-					</form>
-				</CardContent>
-			</Card>
-		</div>
-	);
+export default async function SignInPage() {
+	const db = await getDbStatus();
+	return <SignInForm dbReady={db.ready} dbError={db.error} />;
 }
